+++ src/styled_char.rs
@@ -35,10 +35,24 @@ impl StyledChar {
         self.nude_char = nude_char;
         self.styled_char = self.compound_style.apply_to(self.nude_char);
     }
+    pub fn get_char(&self) -> char {
+        self.nude_char
+    }
     /// Change the fg color, keeping the char, bg color and attributes
     pub fn set_fg(&mut self, color: Color) {
         self.compound_style.set_fg(color);
         self.styled_char = self.compound_style.apply_to(self.nude_char);
+    }
+    pub fn get_fg(&self) -> Option<Color> {
+        self.compound_style.get_fg()
+    }
+    /// Change the bg color, keeping the char, fg color and attributes
+    pub fn set_bg(&mut self, color: Color) {
+        self.compound_style.set_bg(color);
+        self.styled_char = self.compound_style.apply_to(self.nude_char);
+    }
+    pub fn get_bg(&self) -> Option<Color> {
+        self.compound_style.get_bg()
     }
     /// Change the style (colors, attributes) of the styled char
     pub fn set_compound_style(&mut self, compound_style: CompoundStyle) {